+++ lib/models/wheel_label_style.dart
@@ -11,9 +11,17 @@ class WheelLabelStyle {
   /// Use this to tweak the alignment (e.g., pi/2 for perpendicular).
   final double angle;
 
+  /// How visual overflow should be handled.
+  final TextOverflow overflow;
+
+  /// An optional maximum number of lines for the text to span, wrapping if necessary.
+  final int? maxLines;
+
   /// Creates a [WheelLabelStyle].
   const WheelLabelStyle({
     this.labelStyle,
     this.angle = 0.0,
+    this.overflow = TextOverflow.clip,
+    this.maxLines = 1,
   });
 }